Publicize: Force theme support for post-thumbnails feature
Forces the "Featured image" UI to appear in the block editor when the
publicize module is enabled, even if the theme doesn't support the
"post-thumbnails" feature.

Posts to Twitter and Facebook use the featured image if it's available,
so even if a theme doesn't support it there's still value in allowing
users to set it.